Move transaction traversal operations from big map operations
 * use the generic TezosOptics sub-objects to navigate the operations'
   structure
